ex00 loading data base next step => parsing input file

+++ ex00/BitcoinExchange.hpp
@@ -33,7 +33,7 @@ class BitcoinExchange {
   
   private :
     std::string csvfile;
-    std::map<double, float> bitcoin;
+    std::map<int, float> bitcoin;
 
 
   public :
